+++ eslint.config.js
@@ -3,7 +3,7 @@ import johnduganConfig from './configs/index.js';
 const config = [
     ...johnduganConfig,
     {
-        ignores: ['node_modules/', 'dist/', 'build/']
+        ignores: ['node_modules/', 'dist/', 'build/', 'tests/fixtures/']
     }
 ];